+++ tests/lint/too-many-parameters.pas
@@ -1,0 +1,17 @@
+unit ManyParams;
+
+interface
+
+implementation
+
+procedure ManyParams(A, B, C, D, E, F, G, H: Integer);
+begin
+  WriteLn(A);
+end;
+
+procedure FewParams(A, B: Integer);
+begin
+  WriteLn(A);
+end;
+
+end.